requirements.txt
django==5.0.5 
djangorestframework==3.15.1
sqlite==3.41.2
